afxdp/cos/tx_completion: relocate inline tests to tx_completion_tests.rs
Same playbook. 228-LOC inline test block extracted.
After: tx_completion.rs drops from 788 → 563 LOC.
cargo build/test 92/878.
